edit entry updated according to designs
