+++ app/javascript/src/components/Invoices/common/NewLineItemRow/NewLineItemStatic.tsx
@@ -1,92 +1,159 @@
-import React, { useState } from "react";
+import React, { useRef, useState, useEffect } from "react";
 
 import dayjs from "dayjs";
-import { currencyFormat, minToHHMM, lineTotalCalc } from "helpers";
-import { DotsThreeVerticalIcon, EditIcon, DeleteIcon } from "miruIcons";
+import { minFromHHMM, minToHHMM, lineTotalCalc, currencyFormat } from "helpers";
+import { DeleteIcon, CalendarIcon } from "miruIcons";
+import TextareaAutosize from "react-autosize-textarea";
 
-const NewLineItemStatic = ({ currency, item, setIsEdit, handleDelete }) => {
-  const [isSideMenuVisible, setIsSideMenuVisible] = useState(false);
-  const [showEditMenu, setShowEditMenu] = useState(false);
-  const hoursLogged = minToHHMM(item.quantity);
-  const date = dayjs(item.date).format("DD.MM.YYYY");
-  const totalRate = lineTotalCalc(item.quantity, item.rate);
-  const name = item.name || `${item.first_name} ${item.last_name}`;
+import CustomDatePicker from "common/CustomDatePicker";
+
+const NewLineItemStatic = ({
+  currency,
+  item,
+  handleDelete,
+  setSelectedOption,
+  selectedOption,
+}) => {
+  const strName = item.name || `${item.first_name} ${item.last_name}`;
+  const [name, setName] = useState<string>(strName);
+  const [lineItemDate, setLineItemDate] = useState(item.date);
+  const [description, setDescription] = useState<string>(item.description);
+  const [rate, setRate] = useState<number>(item.rate);
+  const [quantity, setQuantity] = useState<any>(minToHHMM(item.quantity));
+  const [lineTotal, setLineTotal] = useState<string>(item.lineTotal);
+  const [showDatePicker, setShowDatePicker] = useState<boolean>(false);
+  const [showCalendarIcon, setShowCalendarIcon] = useState<boolean>(false);
+  const datePickerRef = useRef(null);
+
+  useEffect(() => {
+    const names = name.split(" ");
+    const newItem = {
+      ...item,
+      first_name: names.splice(0, 1)[0],
+      last_name: names.join(" "),
+      name,
+      date: lineItemDate,
+      description,
+      rate,
+      quantity: minFromHHMM(quantity),
+      lineTotal,
+    };
+
+    const selectedOptionArr = selectedOption.map(option => {
+      if (
+        (option.id && option.id === item.id) ||
+        (option.timesheet_entry_id &&
+          option.timesheet_entry_id === item.timesheet_entry_id)
+      ) {
+        return newItem;
+      }
+
+      return option;
+    });
+
+    name && setSelectedOption(selectedOptionArr);
+  }, [name, lineItemDate, description, quantity, rate, lineTotal]);
+
+  const closeEditField = event => {
+    if (event.key === "Enter") {
+      event.target.blur();
+    }
+  };
+
+  const handleSetRate = e => {
+    const qtyInMin = Number(minFromHHMM(quantity));
+    setRate(e.target.value);
+    setLineTotal(lineTotalCalc(qtyInMin, e.target.value));
+  };
+
+  const handleSetQuantity = e => {
+    const qtyInMin = Number(minFromHHMM(e.target.value));
+    setQuantity(e.target.value);
+    setLineTotal(lineTotalCalc(qtyInMin, rate));
+  };
+
+  const handleDatePicker = selectedDate => {
+    setLineItemDate(selectedDate);
+    setShowDatePicker(false);
+  };
 
   return (
     <>
-      <tr
-        className="invoice-items-row"
-        onMouseEnter={() => {
-          setIsSideMenuVisible(true);
-        }}
-        onMouseLeave={() => {
-          setIsSideMenuVisible(false);
-          setShowEditMenu(false);
-        }}
-        onMouseOver={() => {
-          setIsSideMenuVisible(true);
-        }}
-      >
+      <tr className="invoice-items-row cursor-pointer">
         <td className="px-1 py-3 text-left text-base font-normal text-miru-dark-purple-1000 ">
-          {name}
+          <input
+            className="focus:outline-none w-full rounded bg-transparent p-1 text-sm font-medium text-miru-dark-purple-1000 focus:border-miru-gray-1000 focus:bg-white focus:ring-1 focus:ring-miru-gray-1000"
+            placeholder={name}
+            type="text"
+            value={name}
+            onChange={e => setName(e.target.value)}
+            onKeyDown={closeEditField}
+          />
         </td>
-        <td className="px-1 py-3 text-right text-base font-normal text-miru-dark-purple-1000 ">
-          {date}
-        </td>
-        <td className="px-1 py-3 text-right text-base font-normal text-miru-dark-purple-1000 ">
-          {currencyFormat(currency, item.rate)}
-        </td>
-        <td className="px-1 py-3 text-right text-base font-normal text-miru-dark-purple-1000 ">
-          {hoursLogged}
-        </td>
-        <td className="px-1 py-3 text-right text-base font-normal text-miru-dark-purple-1000 ">
-          {currencyFormat(currency, totalRate)}
-        </td>
-        <td className="w-10">
-          {isSideMenuVisible && (
-            <div className="relative">
-              <button
-                className="m-0.5 flex h-9 w-9 items-center justify-center rounded bg-miru-gray-1000"
-                onClick={() => {
-                  setShowEditMenu(true);
-                }}
-              >
-                <DotsThreeVerticalIcon
-                  color="#000000"
-                  size={16}
-                  weight="bold"
-                />
-              </button>
-              {showEditMenu && (
-                <ul className="absolute right-0 top-10 w-30 rounded bg-white shadow-c1">
-                  <li>
-                    <button
-                      className="flex w-full items-center px-2.5 py-4 text-left hover:bg-miru-gray-100"
-                      onClick={() => {
-                        setIsEdit(true);
-                      }}
-                    >
-                      <EditIcon color="#5b34ea" size={16} weight="bold" />
-                      <span className="ml-2 text-miru-han-purple-1000">
-                        Edit
-                      </span>
-                    </button>
-                  </li>
-                  <li>
-                    <button
-                      className="flex w-full items-center px-2.5 py-4 text-left hover:bg-miru-gray-100"
-                      onClick={() => {
-                        handleDelete(item);
-                      }}
-                    >
-                      <DeleteIcon color="#E04646" size={16} weight="bold" />
-                      <span className="ml-2 text-miru-red-400">Delete</span>
-                    </button>
-                  </li>
-                </ul>
-              )}
+        <td className="relative px-1 py-3 text-right text-base font-normal text-miru-dark-purple-1000 ">
+          <div onClick={() => setShowDatePicker(!showDatePicker)}>
+            <input
+              placeholder="Select Date"
+              type="text"
+              value={lineItemDate && dayjs(lineItemDate).format("DD.MM.YYYY")}
+              className={`focus:outline-none w-full cursor-pointer appearance-none rounded border-0 border-transparent bg-transparent p-1 pl-2 text-right text-sm font-medium text-miru-dark-purple-1000 focus:bg-white focus:bg-white focus:ring-1 focus:ring-miru-gray-1000 ${
+                showCalendarIcon ? "pr-9" : "pr-1"
+              }`}
+              onBlur={() => setShowCalendarIcon(false)}
+              onFocus={() => setShowCalendarIcon(true)}
+              onKeyDown={closeEditField}
+            />
+            {showCalendarIcon && (
+              <CalendarIcon
+                className="absolute top-0 right-0 mx-3 mt-4"
+                color="#5B34EA"
+                size={20}
+              />
+            )}
+          </div>
+          {showDatePicker && (
+            <div className="absolute" ref={datePickerRef}>
+              <CustomDatePicker
+                date={new Date(lineItemDate)}
+                handleChange={handleDatePicker}
+                setVisibility={setShowDatePicker}
+                wrapperRef={datePickerRef}
+              />
             </div>
           )}
+        </td>
+        <td className="px-1 py-3 text-right text-base font-normal text-miru-dark-purple-1000 ">
+          <input
+            className="focus:outline-none w-full rounded bg-transparent p-1 text-right text-sm font-medium text-miru-dark-purple-1000 focus:border-miru-gray-1000 focus:bg-white focus:ring-1 focus:ring-miru-gray-1000"
+            placeholder="Rate"
+            type="text"
+            value={rate}
+            onChange={handleSetRate}
+            onKeyDown={closeEditField}
+          />
+        </td>
+        <td className="px-1 py-3 text-right text-base font-normal text-miru-dark-purple-1000 ">
+          <input
+            className="focus:outline-none w-full rounded bg-transparent p-1 text-right text-sm font-medium text-miru-dark-purple-1000 focus:border-miru-gray-1000 focus:bg-white focus:ring-1 focus:ring-miru-gray-1000"
+            placeholder="Quantity"
+            type="text"
+            value={quantity}
+            onChange={handleSetQuantity}
+            onKeyDown={closeEditField}
+          />
+        </td>
+        <td className="px-1 py-3 text-right text-base font-normal text-miru-dark-purple-1000 ">
+          {currencyFormat(currency, lineTotal)}
+        </td>
+        <td className="w-10">
+          <button
+            className="flex w-full items-center rounded p-2.5 text-center hover:bg-miru-gray-200"
+            onClick={() => {
+              handleDelete(item);
+            }}
+          >
+            <DeleteIcon color="#5B34EA" size={16} weight="bold" />
+          </button>
         </td>
       </tr>
       <tr>
@@ -94,7 +161,14 @@ const NewLineItemStatic = ({ currency, item, setIsEdit, handleDelete }) => {
           className="border-b-2 border-miru-gray-200 px-1 pb-4 text-left text-xs font-normal text-miru-dark-purple-400"
           colSpan={2}
         >
-          {item.description}
+          <TextareaAutosize
+            className="focus:outline-none w-full rounded bg-transparent p-1 text-sm font-medium text-miru-dark-purple-1000 focus:border-miru-gray-1000 focus:bg-white focus:ring-1 focus:ring-miru-gray-1000"
+            placeholder="Description"
+            type="text"
+            value={description}
+            onChange={e => setDescription(e.target["value"])}
+            onKeyDown={closeEditField}
+          />
         </td>
         <td className="border-b-2 border-miru-gray-200" colSpan={3} />
       </tr>
